Pull name importing into `ImportModel.h`
